requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-extensions==2.1.5
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-multiple-models==2.1.3
django-sendgrid-v5==0.7.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
future==0.17.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
oauthlib==2.1.0
Pillow==5.4.1
psycopg2==2.7.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
PyPDF2==1.26.0
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.9
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.1.0
sendgrid==5.6.0
six==1.12.0
txaio==18.8.1
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
zope.interface==4.6.0
